fix for running on raspberry pi and in systemd
Raspberry pi 1 model B is a 32-bit system, so %lu is not the right
formatting for size_t. `%zu` is the proper way to print a size_t on
any system.
source: https://stackoverflow.com/a/940119
source: man 3 printf

systemd/journald is not terminal output, so C automatically sets
stdout to block buffered instead of line buffered. We need
`setlinebuf(stdout)` to make sure it stays line buffered.